+++ resources/js/app.ts
@@ -6,15 +6,15 @@ import './globals/sideover';
 import './globals/neura-kit';
 import './globals/command';
 import './globals/context-menu';
+import './globals/spotlight';
+import './globals/clipboard';
 import './globals/color-picker';
 import './globals/dropzone';
 import './globals/chart';
 import './globals/lottie';
 import './globals/editor';
-import './globals/clipboard';
 import './globals/phone-input';
 import './globals/tree';
 import './globals/flow';
-import './globals/spotlight';
 import './components';